gen-ai/tutorials/generation/deep_researcher/pyproject.toml
[project]
name = "deep-researcher"
version = "0.0.0"
description = "Deep researcher usage examples"
requires-python = ">=3.11,<3.14"
readme = "README.md"
dependencies = [
    "gllm-core>=0.4.0,<0.5.0",
    "gllm-inference>=0.5.0,<0.6.0",
    "gllm-generation[google,openai,parallel,perplexity]>=0.5.0,<0.6.0",
    "python-dotenv>=1.0.0,<2.0.0",
]

[[tool.uv.index]]
name = "gen-ai-internal"
url = "https://glsdk.gdplabs.id/gen-ai-internal/simple/"

[tool.uv.sources]
gllm-core = { index = "gen-ai-internal" }
gllm-inference = { index = "gen-ai-internal" }
gllm-generation = { index = "gen-ai-internal" }
